requirements.txt
# requirements.txt

fastapi
uvicorn
python-dotenv
chromadb
google-genai
langdetect